The Human Resources Project was developed on Azure Devops with agile scrum methodology. In addition, in this project, it was developed on onion layer architecture and SOLID principles were considered. This project has authentication and authorization to log in as company manager and personal roles. Users whose role is the company manager can access all informations and expenses of the personnel, approve or reject their expenses.
The Human Resources Project was developed on Azure Devops with agile scrum methodology. In addition, in this project, it was developed on onion layer architecture and SOLID principles were considered.
This project has authentication and authorization to log in as company manager and personal roles. Users whose role is the company manager can access all informations and expenses of the personnel, approve or reject their expenses.